+++ src/clj/l22/routes/home.clj
@@ -10,8 +10,8 @@
    [l22.profile :refer [profile-login profile-show]]
    [ring.util.response]))
 
-(def ^:private version "0.12.18")
-(def ^:private updated_at "2023-04-17 22:42:18")
+(def ^:private version "0.14.23")
+(def ^:private updated_at "2023-05-31 21:23:12")
 
 (defn home-page [{:keys [flash] :as request}]
   (layout/render request "home.html"